+++ NEWS.md
@@ -18,6 +18,8 @@
 * New [`Reactable.onStateChange()`](https://glin.github.io/reactable/articles/javascript-api.html#reactable-onstatechange)
   method in the JavaScript API that sets up a function to be called whenever the
   table state changes ([#265](https://github.com/glin/reactable/issues/265)).
+* New [`Reactable.setData()`](https://glin.github.io/reactable/articles/javascript-api.html#reactable-setdata)
+  method in the JavaScript API for updating the table data ([#278](https://github.com/glin/reactable/issues/278)).
 * `getReactableState()` now includes the current sorted columns ([#265](https://github.com/glin/reactable/issues/265)).
 
 ## Minor improvements and bug fixes